Create data cleaningRestAPI
This Python script is an xApp fetches cell report data from an InfluxDB RESTful API, processes it into a structured format, and saves it as CSV files in a folder called "CELL_REPORT" to be used for ML methods. It fetched data of DB of TeraVM RIC TESTER, through an xApp. This is a part of project entitled "REACH" was done by Vahid Kouh Daragh as a research fellow for University of york.